backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
langchain
langchain-openai
chromadb
pydantic
